39 55 fixes
