+++ src/RudolfReducer.ts
@@ -55,7 +55,6 @@ export class RudolfReducer extends ImmerReducer<RudolfStore> {
   createTree(premiseArray: string[]) {
     this.draftState.tree = parsePremises(premiseArray)
     this.draftState.nextRow = premiseArray.length + 1
-    //this.draftState.justifications = { 1: { rule: '', parentRow: '' } }
     this.draftState.justifications = {}
   }
 
@@ -119,7 +118,6 @@ const premiseArray = initialPremises.split(',')
 export const initialState: RudolfStore = {
   tree: parsePremises(premiseArray),
   nextRow: premiseArray.length + 1,
-  //justifications: { 1: { rule: 'AS', parentRow: '' } },
   justifications: {},
   feedback: { errorMessage: 'Nothing yet.' },
 }
